This moves methods that were in the `works_controller` that did whitespacing into a new `Tufts::Normalizer` module.
That module is used in `works_controller` and in the `ImportRecord` class.

It noramlizes the field contents in `ImportRecord` in the same way that is being done in `works_controller`

I also updated the sample XML file with whitespace and the `abstract` and `description` elements. This lets me
test that those fields are being whitespaced correctly.
