+++ spec/fixtures/files/mira_xml.xml
@@ -21,8 +21,16 @@ http://www.openarchives.org/OAI/2.0/OAI-PMH.xsd">
           <tufts:filename>pdf-sample.pdf</tufts:filename>
           <tufts:visibility>open</tufts:visibility>
           <model:hasModel>Pdf</model:hasModel>
-          <dc:title>President Jean Mayer speaking at commencement, 1987</dc:title>
-          <dc:description>5x7</dc:description>
+          <dc:title>President Jean Mayer speaking
+          at commencement, 1987</dc:title>
+          <dc11:description>
+            A long       description             with 
+                           plenty
+          of          whitespace</dc11:description>
+          <dc:abstract>
+            Another long       description             with 
+                           plenty
+          of          whitespace</dc:abstract>
           <dc:source>UA136</dc:source>
           <dc:coverage>Medford;7014023;42.417;-71.100</dc:coverage>
           <dc:temporal>1987</dc:temporal>
