+++ spec/services/tufts/import_service_spec.rb
@@ -31,7 +31,7 @@ describe Tufts::ImportService, :workflow do
     end
 
     it 'has the title from the imported record' do
-      title = 'President Jean Mayer speaking at commencement, 1987'
+      title = "President Jean Mayer speaking\n          at commencement, 1987"
 
       expect(service.import_object!).to have_attributes(title: [title])
     end
